made auth options exportable
