+++ app/app/api/pin/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { createClient } from '@/utils/supabase/server';
 import { getServerSession } from 'next-auth';
-import authOptions from '../auth/[...nextauth]/route';
+import { authOptions } from "@/app/lib/auth";
 
 export async function POST(request: NextRequest) {
     try {
